pyproject.toml
[tool.poetry]
name = "test_task"
version = "0.1.0"
description = "AI-Enhanced Notes Management System"
authors = ["Dmytro-Kruhlov"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
pydantic = {extras = ["email"], version = "^2.10.6"}
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.0"
nltk = "^3.8.1"
alembic = "^1.13.1"
passlib = "^1.7.4"
redis = "^5.2.1"
google-generativeai = "^0.8.4"
psycopg2 = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
psycopg2-binary = "^2.9.10"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
fastapi-jwt-auth = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.26.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
